Java 5 - Remove unnecessary unboxing. (#3483)

+++ src/net/sourceforge/kolmafia/request/ArcadeRequest.java
@@ -252,7 +252,7 @@ public class ArcadeRequest extends GenericRequest {
   private static String findChoiceName(final int index) {
     if (indices != null && choices != null) {
       for (int i = 0; i < indices.length; ++i) {
-        if (indices[i].intValue() == index) {
+        if (indices[i] == index) {
           return choices[i];
         }
       }
